Fix invocation of normalize_elf_addr()
In our normalization logic we iterate over addresses and proc maps
entries. When we have an address that is before the entry's range, we
mark said address as unknown. When the address is past the entry's
range, we advance to the next entry. However, we once we advanced to the
next entry, we do not check whether this entry actually covers the
address -- both its start and end address could be past said address. If
that were the case, we'd end up invoking normalize_elf_addr() and can
trigger an underflow in there.
Fix this problem by reordering the range checks: first advance entries
while the address is past each entry's end address and then check if the
start address is before it.

Signed-off-by: Daniel Müller <deso@posteo.net>