Remove 'symbolize' feature
With everything requiring the regex crate removed, we are now in a
position to tremendously simplify our conditional compilation logic.
Specifically, we can remove the symbolize feature altogether and replace
all the conditional compilation it enabled with unconditionally compiled
code. Because we still want to make the lru dependency optional, just
conditionally the ELF cache functionality itself on it.

Signed-off-by: Daniel Müller <deso@posteo.net>