Remove find_addresses_opt() functionality
This change removes the find_addresses_opt() functionality from the
crate. We are in the process of reworking everything related to finding
addresses for symbols, and the additional feature/option support needs
to be part of that. Given that it's not actually clear that we need
the advanced filtering, remove these bits for the time being and until a
proven use case has been found. At that point we can add it back with
proper tests.

Signed-off-by: Daniel Müller <deso@posteo.net>