Move blazesym_find_addrs() into inspect module
This change moves the blazesym_find_addrs() function along with a bunch
of type definitions into the newly created 'inspect' sub-module.
Upcoming changes will adapt it to work with the inspector type instead
of on a symbolizer.
I renamed functions & types in the process of the move to reflect new
naming (e.g., blaze_ prefix as opposed to blazesym_, shorter names with
less duplication, etc.).

Signed-off-by: Daniel Müller <deso@posteo.net>