Use nightly Rust for code coverage job
We missed explicitly installing the most recent Rust for the coverage
job. That resulted in us using the Ubuntu default Rust toolchain.
Especially for code coverage, it can pay off to use the most recent Rust
version, which uses more up-to-date LLVM toolchains, which in turn may
provide better coverage instrumentation. In fact, code coverage still
works best on nightly, so switch to using a nightly toolchain for this
job.

Signed-off-by: Daniel Müller <deso@posteo.net>