Introduce normalize::Handler trait to better separate concerns
The address normalization logic mixes two concerns in the main
algorithm: the matching up of proc maps entries and input addresses as
well as implementation details on what to do with the resulting
<address, entry> pairs.
With this change we introduce a trait that helps us isolate the two
concerns better. Specifically, the core algorithm in
normalize_sorted_user_addrs_with_entries() stays there, while the
normalization specific logic is moved into the NormalizationHandler
type, which implements the newly introduced Handler trait.

Signed-off-by: Daniel Müller <deso@posteo.net>