Introduce with_ordered_elems() helper function
User address normalization comes in two variants: one that expects a
sorted array of address and one that doesn't. The latter makes sure that
we sort the input first into a temporary allocated array, perform the
normalization, and then rearrange the result in such a way that it
matches the ordering of the unsorted input.
With upcoming changes we need similar functionality of sorting input,
invoking a function, and then restoring the original ordering in more
places. As such, this change factors out a utility function that takes
care of that in a generic manner.

Signed-off-by: Daniel Müller <deso@posteo.net>