Use internal APIs in process symbolization
So far we were using the public normalization APIs as part of the
process address symbolization procedure. This has several drawbacks:
- we read build IDs of the various ELF files involved, but never use
  them
- we work with symbolic maps entry paths as opposed to map files, which
  can cause problems with file system namespaces, for example
- we allocate memory for all the meta data that we gather and pass
  around

In short, we do a bunch of work that makes sense in the public
normalization APIs, but is entirely unnecessary when the result is used
only for symbolization itself.
With this change we rework this logic, using the previously introduced
normalize::Handler trait to tailor behavior to the symbolization use
case, and fix the aforementioned problems.

Signed-off-by: Daniel Müller <deso@posteo.net>