Rename Binary meta data variant to Elf
This change renames the Binary meta data variant to Elf. For the time
being we decided to err on the side of having more specific variants,
which may be easier to extend in a sensible way in the future (e.g., if
future supported binary formats don't have a build ID it's more natural
to have a new variant as opposed to fudging everything as a "binary" and
then having optional fields).

Signed-off-by: Daniel Müller <deso@posteo.net>