Stop using Git LFS for large benchmark files
Git LFS turned out to be not exactly suitable for our intents and
purposes. Most importantly, GitHub's LFS imposes a fixed,
non-replenishable bandwidth limit (which includes downloads...) on the
uploader. Second, because of the way it works, once installed on the
system, by default downloading will happen as part of clone/checkout.
Neither "qualities" are particularly appealing to us.
With this change we move large binary blobs away from LFS and into a
separate repository. Because GitHub has a 100 MiB file size limit, we
split files accordingly. Our build script contains the logic for
downloading all parts and piecing them back together. All of this is
controlled by the generate-bench-files features, unchanged.

Signed-off-by: Daniel Müller <deso@posteo.net>