Add unused member to blaze_user_addr_meta_unknown struct
Apparently it is *illegal* to have an empty struct in C. gcc seems to
accept that as an extension of sorts [0]. Be that as it may, during
integration into a C++ code base we noticed an error showing up:
  > empty struct has size 0 in C, size 1 in C++ [-Werror,-Wextern-c-compat]

While we cannot, for the love of it, reproduce that problem in an
isolated manner, make sure to add an unused member to the type
in questions. Also compile a C++ program, not that it would have caught
this issue, but it's probably better than nothing.

[0] https://gcc.gnu.org/onlinedocs/gcc/Empty-Structures.html

Signed-off-by: Daniel Müller <deso@posteo.net>