Make blaze_inspect_elf_src::path a *const _
The blaze_inspect_elf_src::path member currently can be mutated by C
code. That's not exactly the intention. Because it is mutable it also
prevents legitimate assignments of a *const _.
The reason for it being mutable in the first place was mostly because
that's what Rust APIs used in the implementation work with. However, we
should not expose this property.
To that end, this change adjusts the type from *mut _ to *const _.

Signed-off-by: Daniel Müller <deso@posteo.net>