Adjust return type of ElfParser::find_section()
When ElfParser::find_section() cannot find a section, it returns an
error. It is perfectly fine for some ELF sections to not exist and it is
ultimately up to the caller to decide what the higher level behavior
should be. Currently, however, that is simply not done, because the
information that a section does not exist is not first class
information, but clubbed together with other runtime errors that are
generally unrecoverable.
This change fixes this issue by adjusting the method's signature to
return an optional index instead. In so doing we fix a subtle bug
causing the DwarfResolver::find_addr() function to error out in some
cases, when an address could not be found in ELF symbols, instead of
actually consulting DWARF debug information. A test will be added
subsequently.

Signed-off-by: Daniel Müller <deso@posteo.net>