Introduce name index to DebugSyms
With upcoming changes we want to look up DWARF symbols by address.
Currently we only support look up by name. To enable lookup by both we
introduce an additional index that allows for quick search by name,
while sorting the list of symbols by address.
We could opt for adding an index for lookup by address instead, but
lookup via an index induces an additional memory access and will be
slightly penalized in terms of performance. We aim for making
symbolization as fast as possible and so lookup by name is managed
via the index instead.

Signed-off-by: Daniel Müller <deso@posteo.net>