Move ELF address lookup logic out of DwarfResolver
It is confusing that a DwarfResolver favors lookup of addresses from ELF
instead of DWARF information. Not only does it go against what its name
implies, even conceptually the DwarfResolver is part of an ElfResolver
and so we can end up with cyclic logic (at least conceptually) if we
call back up (that's more a theoretical possibility than a real one
right now, because the DwarfResolver only has an ElfParser, but the
larger argument stands).
Move this conditional logic of first attempting lookup via ELF and then
falling back to DWARF from DwarfResolver into ElfResolver.

Signed-off-by: Daniel Müller <deso@posteo.net>