Use lowest match search in ElfParser::find_sym()
The ElfParser::find_sym() method uses rather convoluted and untested
search_address_opt_key() function to search for a symbol. We should be
able to accomplish the same result by just using our lowest match search
to find the first symbol with a given address, and then perform a linear
scan to eliminate candidates that don't fit the expected for type.

Signed-off-by: Daniel Müller <deso@posteo.net>