Stop working on ELF files with ELF symbols in DWARF tests
Now that we support using DWARF when mapping addresses to symbols, get
rid of the "most-DWARF" file we used for testing. Everything is now
using the DWARF-only variant, which makes testing more predictable in
the sense that we would fail if we were to (wrongly) assume ELF symbols
to be present.

Signed-off-by: Daniel Müller <deso@posteo.net>