Introduce ApkElf meta data
With upcoming changes we will add support for normalizing addresses
belonging to ELF binaries that reside in APKs to the crate. Such
addresses require additional meta data as output of the address
normalization process, which includes the APK path.
This change introduces a new meta data variant, ApkElf, that we will use
for representing such addresses.

Signed-off-by: Daniel Müller <deso@posteo.net>