Normalize APK ELF addresses relative to ELF file
We want APK ELF addresses to be normalized relative to the ELF file, as
that allows for easy symbolization if the ELF file is available. The
current code, however, report an address normalized to the containing
APK.
This change adjusts the logic and documents the property. We also start
adding more doc tests.

Signed-off-by: Daniel Müller <deso@posteo.net>