Provide custom Debug impl for PathMapsEntry
We may end up including PathMapsEntry objects in traces. Given that they
contain addresses and other data is not hard to understand when
formatted as decimal values (the default), this change introduces a
custom Debug impl that make sure to print said fields formatted in the
easiest to understand way.

Refs: #223

Signed-off-by: Daniel Müller <deso@posteo.net>