Add addr2line-based DWARF parsing logic
When we originally switched over to a gimli based DWARF parser we were
mostly concerned with it passing all tests while making it fit the
existing interface -- both internally and externally. E.g., we kept most
of the structure inside the DWARF resolver and parser, such as having
separate paths for parsing symbols and parsing line information or
parsing symbols of all units in one go.
Since then we have gotten a better understanding of the DWARF format
itself, its structure and peculiarities and it became more clear that
preserving this internal structure will be limiting in some aspects and
simply be unfeasible in others. E.g., once we start parsing multi-GiB
DWARF files we are likely to run into memory issues with a simple
all-or-nothing approach. With inlined functions this problem will be
immediate, if we were to follow the same approach.
As it turns out the addr2line crate [0] has a battle tested (it seems to
be what is used by Rust's standard library itself) DWARF parser which
already deals with a bunch of quirks found in the wild that is
structured in a way that makes it a good candidate to base our
implementation on. Specifically, its core logic is lazy up to the
individual function level, look up of multiple entities is typically
iterator based (as opposed to allocating a Vec), it has split-DWARF
support, and in general it can be made to provide all the data that we
would like to expose.
With this change we pull in the relevant bits into our dwarf module.
Structure and style is kept roughly the same, but a few changes had to
be made (non-exhaustive list):
- we work with our R typedef (gimli::EndianSlice) instead of being
  generic over any gimli::Reader
  - on top of likely being faster, that will allow a bunch of operations
    to be infallible (e.g., slice extraction)
- the source location logic was adjusted to avoid allocations for paths
  unless necessary
- unnecessary functionality, such as range based lookups, was removed
- new methods were added to support our existing "inspect" functionality
  (currently: lookup by name)
  - right now this is pretty inefficient, as it means lazily loading and
    scanning *all* functions in all units and we do not keep an index
    around (though we could)
- removed the Context type in favor of Units
- removed all split-DWARF support
  - focus is on fixing bugs in existing functionality and getting to
    feature parity

This new functionality is still unused. Subsequent changes will switch
our DwarfResolver over to relying on it.

[0] https://github.com/gimli-rs/addr2line

Signed-off-by: Daniel Müller <deso@posteo.net>