Work with complete Dwarf object when parsing symbols
When we implemented the gimli based DWARF symbol parsing, we opted for
an approach that makes section usage obvious, by loading all possibly
required sections and providing contents to functions as necessary.
However, we won't be able to keep this up. DWARF is just too
intertwined, where everything can reference everything and in the end
we will need a god-object as provided with gimli::Dwarf. We can already
see that if we look at gimli::Attribute::string_value() [0] and the
warning there, which recommends usage of Dwarf::attr_string instead:

> this function does not handle all possible string forms. Use
> Dwarf::attr_string instead.

So switch over to using a gimli::Dwarf object instead. Performance
change seems to be well within noise levels.

[0] https://docs.rs/gimli/0.27.2/gimli/read/struct.Attribute.html#method.string_value

Signed-off-by: Daniel Müller <deso@posteo.net>