Run all benchmarks in CI
In the past we used Git LFS for retrieving large benchmark files and we
didn't want to waste upload/download bandwidth on every CI run. In
addition, we never intended to rely on CI benchmarks, for they are
running in a virtualized and generally unreliable (in terms of
consistent performance) environment.
Despite that, given that by now we switched to pulling said files from a
GitHub hosted Git repository with commit 80f35dada75b ("Stop using Git
LFS for large benchmark files"), it is still somewhat interesting to
have all benchmarks run in CI, because even if the numbers are not
particularly, it could be used to identify trends over time or to notice
outrageous regressions more timely.
With this change we enable running of all benchmarks in CI. We cache all
Rust build artifacts for this job (and only this job), to prevent it
from becoming the longest running part of the overall workflow.

Signed-off-by: Daniel Müller <deso@posteo.net>