Introduce symbolize::GsymData
This change extends the symbolize::Source::Gsym variant by allowing for
construction with user provided data, in addition to the possibility of
providing a path to a Gsym file to load. This is useful in context where
file system involvement is not desired or in which information about the
source file is not available.

Signed-off-by: Daniel Müller <deso@posteo.net>