Remove SymResolver::get_obj_file_name() method
We don't really want to require association of a file name with a
resolver. It is entirely reasonable to have a resolver that just works
on allocated/mmap'ed data that was just handed to us and for which we
don't (and in the general case: cannot) know what file it belongs to.

We don't really need the method. Inside the Inspector we can just use
the path being provided and used for instantiation of the corresponding
resolvers, and elsewhere we just use the file name to spice up the debug
representation or similar. Remove it.

Signed-off-by: Daniel Müller <deso@posteo.net>