Adjust DwarfResolver to use addr2line based functionality
This change switches the DwarfResolver over to using the previously
introduced new DWARF parsing functionality, which is based on code as
found in gimli's addr2line crate.
The new functionality has different performance characteristics, as
follows:

> Benchmarking main/inspect :: lookup_dwarf: Warming up for 5.0000 s
> main/inspect :: lookup_dwarf
>                         time:   [105.28 ms 105.67 ms 106.08 ms]
>                         change: [+196.68% +197.73% +198.79%] (p = 0.00 < 0.02)
>                         Performance has regressed.
>
> main/symbolize :: symbolize_process
>                         time:   [2.1725 ms 2.1862 ms 2.2031 ms]
>                         change: [-80.967% -80.848% -80.661%] (p = 0.00 < 0.02)
>                         Performance has improved.
>
> Benchmarking main/symbolize :: symbolize_dwarf: Warming up for 5.0000 s
> main/symbolize :: symbolize_dwarf
>                         time:   [76.375 ms 76.590 ms 76.884 ms]
>                         change: [-88.845% -88.816% -88.771%] (p = 0.00 < 0.02)
>                         Performance has improved.

Symbolization performance has improved quite a bit, which can likely be
attributed to the more sophisticated attribute tree traversal. The
degradation in "inspect" performance is due to our current brute-force
implementation of "lookup by name". We should be able to improve this
moving forward, but we don't consider the inspection functionality
performance critical and so it has not been further optimized.

Signed-off-by: Daniel Müller <deso@posteo.net>