Publicize a couple of C API struct members
It's conceptually fine for members of struct types exposed via the C API
to be private -- bindgen picks them up no matter what and C doesn't
care. However, if they are private they don't show up in the
documentation, which is not a great experience.
Make all C API struct members public to expose them in the
documentation. Unfortunately there does not seem to exist any flag to
raise an error on the cbindgen side for non-public members, so I suppose
this whack-a-mole is all we can do.

Signed-off-by: Daniel Müller <deso@posteo.net>