Introduce ELF Backend type
For implementation peculiarities on Linux, we cannot make do with an ELF
parser that unconditionally relies on memory mapped data. Rather, for
certain /proc files we are actually forced to use file I/O instead. In
order to switch between the two, introduce the Backend trait that
abstracts over the differences between the two. In a nutshell, for
the mmap backend we continue reading references to objects while for the
file I/O based one we will work with owned objects instead.

Signed-off-by: Daniel Müller <deso@posteo.net>