Convert public APIs to use new Error type
This change switches the public parts of the normalize, symbolize, and
inspect modules over to using the newly introduced Error type. This is
not a wholesale conversion by any means. The intent is to convert more
pieces over over time and hopefully eliminate a bunch of boilerplate in
the process.

Signed-off-by: Daniel Müller <deso@posteo.net>