Adjust ErrorImpl::Dwarf variant
When we originally devised the ErrorImpl::Dwarf variant we had it store
a gimli::Error. But then we changed it to just a mere error string,
thinking that such an error alone would not be very useful. However,
that's not exactly true: while for some variant of this error a proper
crafted string may be more useful, in many of them the description is
not bad.
As such, this change switches the ErrorImpl::Dwarf variant back to
storing a gimli::Error object.

Signed-off-by: Daniel Müller <deso@posteo.net>