Improve Cow handling in ErrorExt trait methods
We decided to special case the &'static str case of context for our
Error type. To that end, in its initial version, the ErrorExt trait just
assumed that context() would always just accept such a static string,
while with_context() was more flexible but would *always* allocate.
With this change we list this restriction by introducing a trait,
IntoCowStr (similar to ToString) that is used by both methods.

Signed-off-by: Daniel Müller <deso@posteo.net>