Rename generate-bench-files feature to generate-large-test-files
With upcoming changes we would like to use the kernel image that we have
used only for benchmarking so far for regression testing as well. As
such, the generate-bench-files switch controlling the download of said
kernel image and generation of derived files is no longer aptly named.
Rename it to generate-large-test-files in anticipation of the new usage.

Signed-off-by: Daniel Müller <deso@posteo.net>