Decrease benchmark runtime
Our integration style benchmarks take quite a while to run. A big reason
for that is that we set the desired sample size and have a rather large
warmup time. We set those parameters to have more stable numbers, and
while that may have a small effect, ultimately the long runtime is a
steep price to pay.
With this change we set a warmup time of 1s and use criterion's default
sample size (100).

Signed-off-by: Daniel Müller <deso@posteo.net>