Properly align benchmark output
Unfortunately criterion with the bencher output format does not align
benchmark results properly ([0]). As a result it is quite hard to read
the summary from a glimpse.
To work around this problem, this change "manually" adjusts the names
of our benchmark functions so that the corresponding results in the
summary report are aligned properly.

[0]: https://github.com/bheisler/criterion.rs/issues/704

Signed-off-by: Daniel Müller <deso@posteo.net>