Hook up demangle flag
With upcoming changes we will add demangling support to the crate. Not
everyone may need or want such support and as such it makes sense to
make it configurable.
This change extends our symbolize::Builder with a flag determining
whether the user desires automatic demangling of symbols or not. The
corresponding C API is adjusted accordingly.

Signed-off-by: Daniel Müller <deso@posteo.net>