Represent lines as u32 and columns as u16
With this change we represent source location lines as u32 and columns
as u16. Doing so is less wasteful than using machine words for both and
it is what both our GSYM and DWARF resolvers use internally. On top of
that doing so provides us with a few padding bytes inside the
symbolize::Sym type that we can use for other purposes, should the need
arise.

Signed-off-by: Daniel Müller <deso@posteo.net>