Introduce SrcLang enum
In order to perform demangling, it can come in useful to know the source
language from which the symbol originates. If known, we can use the
appropriate demangling scheme.
Currently only DWARF carries such language information, but we don't
want to splatter DWARF details throughout the code base. This change
introduces the SrcLang enum that represents the source code language, if
known.

Signed-off-by: Daniel Müller <deso@posteo.net>