Add support for demangling symbols
This change adds support for transparent demangling of symbols.
Demangling is generally runtime configurable (via the previously added
Builder flag), but because it involves additional dependencies there is
also the option to disable it at compile time. If disabled at compile
time, the runtime option becomes a no-op.
Demangling currently happens in a single location inside the
Symbolizer type. This has the advantage of concentrating the feature's
implementation. However, it means that we need to bubble up the inferred
language from all resolvers -- something that the recently added IntSym
type helps us with.
Note that this change only provides *de*mangling support. That is, it
hooks into the APIs mapping addresses to human readable symbols. The
inverse, mapping human readable and potentially *un*mangled symbols to
their mangled counterparts and then to addresses is not currently being
done (that would be a useful addition to the inspect module, but is out
of the scope of this change).

Closes: #50

Signed-off-by: Daniel Müller <deso@posteo.net>