Introduce blazecli
This change introduces a first version of blazecli, a command line
interface to blazesym. The aim of the program is to showcase blazesym's
functionality and help with debugging of issues. Its command interface
strives to stay close to blazesym's public API interface. Currently it
only supports symbolization of addresses inside a process, but it can be
extended as necessary.
The program comes as its own workspace member to keep dependencies
independent of the main library and allow for separate publishing.

  $ cargo run -p blazecli -- symbolize process --pid=765676 0x000000000040110a 0x00007f0264e29510
  > 0x40110a: main@0x40110a+4 :0
  > 0x7f0264e29510: __libc_start_call_main@0x7f0264e29510+139648259063936 :0

Signed-off-by: Daniel Müller <deso@posteo.net>