Expose build ID reading functionality publicly
Being able to read build IDs from arbitrary ELF files may be necessary
in order to better work with the results of the normalization APIs, for
example, when attempting to look up a file by the reported build ID. As
such, it makes sense to export this functionality publicly. With this
change we do just that, inside the newly introduced helper module. While
at it, also add a test for md5 build IDs.

Signed-off-by: Daniel Müller <deso@posteo.net>