Fix maps_file path creation for low addresses
As it turns out, when low addresses are reported inside /proc/[pid]/maps
they may get prefixed with '0' characters in the reported range.
However, this range is not actually what can be looked up below
/proc/self/map_files/: there these leading zeroes are removed instead.
So far we deliberately treated the range *as a string* to *not* have to
deal with any mismatches, but this turns out to be the wrong approach.
Hence, with this change we create the maps file path based on the parsed
range, which takes care of removing any leading zeroes.

Signed-off-by: Daniel Müller <deso@posteo.net>