Introduce symbolize::Sym::offset member
The documentation of the symbolize::Sym::addr member is wrong: the
reported address is pretty much never that of the target process, for
the simple reason that there may not be a target process at all (when
symbolizing an address inside an ELF file, for example). This change
fixes the documentation.
Furthermore, a common use case when symbolizing stack traces, for
example, is to include the offset of the instruction included in the
trace in the symbolized result. Currently that is done by subtracting
the input address from the "symbolized" address. However, this
calculation is fundamentally broken most of the time in a process
context, because process addresses are not normalized addresses, but our
Symbolizer reports normalized addresses. Hence, when relocations are or
randomizations are applied, the result of the operation is meaningless.
To support such use cases, this change adds an explicit offset member to
the symbolize::Sym type. As the name tries to convey, this member holds
the byte offset of the symbolized instruction from the start of the
function. Furthermore, using this member it is possible to safely
calculate "process offset" of the address in question:
  <process offset> = <process addr> - (sym.addr + sym.offset)

Signed-off-by: Daniel Müller <deso@posteo.net>