Report ELF symbols with bogus size if only conceivable match
Starting with commit 1b6878918160 ("Use lowest match search in
ElfParser::find_sym()") we honored the symbol's size when attempting
to symbolize an address based on ELF information. However, it turns out
that the size may not actually be reliable information: we have seen
cases where a function was reported as much smaller than the actual code
that it comprised at runtime (the addresses of which could show up in
backtraces). Whether that is due to runtime linker tricks or other
shenanigans, we should be able to handle such cases.
With this change we adjust the ELF parser symbol lookup logic to handle
such cases more gracefully. Specifically, if an address falls between
two symbols such as 0x29d90 in:

  0000000000029d00 <__libc_init_first>
     29d00:       f3 0f 1e fa             endbr64
     29d04:       c3                      ret
     29d05:       66 2e 0f 1f 84 00 00    cs nopw 0x0(%rax,%rax,1)

     [ more code that according to ELF does not belong to __libc_init_first ]

  0000000000029dc0 <__libc_start_main>
     29dc0:       f3 0f 1e fa             endbr64
     29dc4:       41 57                   push   %r15
     29dc6:       49 89 cf                mov    %rcx,%r15

We map it to the only possible match -- __libc_init_first in this case.

Fixes: #269

Signed-off-by: Daniel Müller <deso@posteo.net>