Fix nightly benchmark builds
Commit cdd5b9c1a057 ("Replace env::args().next() with
env::current_exec()") broke the nightly benchmarks and that was not
caught because...GitHub Actions uses some more than questionable
defaults for the default shell, which supposedly is bash, but it is
invoked with fewer error reporting options than if we explicitly asked
for it [0]. Make. It. Stop.
Explicitly specify the shell and fix the build.

[0] https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#jobsjob_idstepsshell

Signed-off-by: Daniel Müller <deso@posteo.net>