Add workflow for building statically linked binary
Sometimes no suitable build environment is at hand but a binary of the
program desired. For such cases it makes sense to have access to a
statically linked binary that essentially is able to run on any Linux.
This change introduces a workflow that can be used to produce such a
binary. It can be triggered on-demand and will produce such a binary and
make it available as a downloadable artifact.

Signed-off-by: Daniel Müller <deso@posteo.net>