Provide custom Debug impl for symbolize::Source
We may end up including symbolize::Source objects in traces. Right now
their Debug representation effectively repeats the name of the inner
type (because that's the name of the variant). E.g.,
  > Elf(Elf("/lib64/libc.so.6"))

While not wrong, it's more information than typically necessary. Shorten
the representation by omitting the variant name and just forwarding to
the inner object.

Refs: #223

Signed-off-by: Daniel Müller <deso@posteo.net>