Unify output of symbolization result in examples
We have five "examples", all of which format their symbolized output
slightly differently. This is the cause of bugs and it's a maintenance
nightmare. These kinds of code snippets also tend to get copied around
everywhere, so it makes sense to make them "nice".
Unify them and shorten them. The produced format is chosen to be
readable while maintaining that the algorithm can be understood and it
only contains a single line formatting branch.

  $ cargo run -p blazecli symbolize elf --path /lib64/libc.so.6 00000000000d93a1
  > 0xd93a1: __readdir_r @ 0xd93a0+0x1 :0

Signed-off-by: Daniel Müller <deso@posteo.net>