Stop returning symbols without name or address
It is believed to be extremely rare that source code information for an
address is available but no symbol name: presumably that would require a
specially crafted DWARF file, if allowed at all.
With this change we ignore such cases and stop returning a Sym object to
begin with. In so doing we can now properly treat the symbol name,
address, and offset as unconditionally present.

Signed-off-by: Daniel Müller <deso@posteo.net>