Make GsymContext::get_str() return OsStr
Adjust the GsymContext::get_str() method to return an OsStr instance on
success. Doing so eliminates a potential error path while minimizing
constraints imposed on the representation. That is to say, given that we
use this method for reading paths, which are later represented as Path,
having a str be the intermediate representation is dangerous, as it has
the potential to wrongly eliminate some valid paths that are not valid
strings.

Signed-off-by: Daniel Müller <deso@posteo.net>