Split symbolize::Sym::path member into dir and file
All our supported formats internally store source code information as a
directory path and a file name. Many clients may reasonably expect this
split and while we could always do it retroactively, it makes sense to
stay as close to the supported formats as possible. Down the line, that
could allow us to stop allocating new buffers altogether and just hand
out references, in a more zero-copy (and zero-alloc) fashion.
With this change we split the symbolize::Sym::path member into dir and
file attributes to represent this split in the library. Note that
currently either both the directory and file are present or neither is.
We could encode this constraint in the type system and it would simplify
client code somewhat, but we opted not to for symmetry with other
members and because it is unclear whether this constraint is truly
guaranteed to true unconditionally in the future.

Signed-off-by: Daniel Müller <deso@posteo.net>