Replace rlib crate-type with lib
We don't really need the library to be an rlib from what I gather. We
need it to be available as:
- a static library (staticlib)
- a dynamic library (cdynlib)
- and as a Rust library

For the latter, lib is the default [0] (which likely is just an alias
for rlib) and I don't think we have any good reason to overwrite this
default. Also sneak in a comment fix while at it.

[0]: https://doc.rust-lang.org/reference/linkage.html

Signed-off-by: Daniel Müller <deso@posteo.net>