Properly match on the GSym address info type
When switching to an iterator based approach in parse_address_data(), we
eliminated one of the type info checks and the printing of a warning if
a type is not recognized.
Bring it back in the caller of the parse_address_data() function and use
a proper match there in anticipation of inline information parsing which
shall be added there eventually.

Signed-off-by: Daniel Müller <deso@posteo.net>