Make GsymResolver::find_line_info() report more error context
We originally introduced find_line_info_impl() in
GsymResolver::find_line_info() to short circuit on the various error
paths, because it was way too cumbersome to report custom errors on
each.
By now we have the means for more concise error chaining and so we
remove this internal function and report more detailed errors for each
of the potential early returns.

Signed-off-by: Daniel Müller <deso@posteo.net>