Use alternate format when formatting hex and bin numbers
As it turns out, Rust can automatically take care of prefixing
hexadecimal numbers with `0x`, as long as we use the alternate format.
This has the advantage that it takes effect for all numbers in a
structured type, e.g., when we print an array of addresses in
hexadecimal format.
With this change we use this alternate format.

Signed-off-by: Daniel Müller <deso@posteo.net>