Fix potentially incorrect reporting of symbols from ELF resolver
When we search for a symbol by address in ELF, we were missing an
additional check that the symbols we iterate over are still within the
desired address range. That's necessary because we effectively filter by
symbol type and such filtering may advance to symbols that are no longer
matching from an address perspective.
This change fixes the problem by adding the necessary address check.

Signed-off-by: Daniel Müller <deso@posteo.net>