Make find_sym() short circuiting
If the find_sym() function does not find a match, internal symbol
iteration will still traverse all symbols past the first candidate we
found. That's unnecessary and can be a source of performance problems.
Make the function short-circuiting such that once we found an address
that is past the input address, we return.

Signed-off-by: Daniel Müller <deso@posteo.net>