Flatten symbolize return type
The doubly nested Vec<Vec<>> construct being returned from the
Symbolizer::symbolize() method has basically been there since the
beginning of time. However, it was also shown to be one of the least
understood aspects of the library for users.
This change reworks the return type. We are now returning a single level
of symbols, but associate each with an index of the input address it
corresponds to. In so doing we can express both the lack of
symbolization as well as the presence of one or more inlined functions
at the address in question.

Signed-off-by: Daniel Müller <deso@posteo.net>