Introduce internal FrameCodeInfo type
Our SymResolver::find_line_info() method currently returns a single
AddrCodeInfo object. That is insufficient when it comes to inline
function information.
Rather than going down the route of returning a vector of those objects,
let's structure things a bit more. With this change we introduce the
FrameCodeInfo type. AddrCodeInfo is reworked to contain such an object
for the function being symbolized, along with an (optional) array
representing inlined functions.
Long-term I hope that we will move to such a scheme for the public
facing API as well, as it is much more telling than returning a mere
array of AddrCodeInfo (or Sym) objects.

Signed-off-by: Daniel Müller <deso@posteo.net>