Change symbol used in ELF demangle test
This patch changes the symbol used for testing purposes in the ELF
demangle test from Normalizer::new to Normalizer::normalize_user_addrs_sorted.
The former has been seen to be reported as Inspector::new on release
builds, which has the exact same body and may be merged as an
optimization.

Signed-off-by: Daniel Müller <deso@posteo.net>